Move logger to module
