+++ test/unit/test_logger.py
@@ -17,8 +17,11 @@ import os
 from tempfile import NamedTemporaryFile
 from unittest import mock, skipIf
 
-from qiskit_ibm_runtime import QISKIT_IBM_RUNTIME_LOG_FILE, QISKIT_IBM_RUNTIME_LOG_LEVEL
-from qiskit_ibm_runtime.utils.utils import setup_logger
+from qiskit_ibm_runtime.utils.logging import (
+    QISKIT_IBM_RUNTIME_LOG_FILE,
+    QISKIT_IBM_RUNTIME_LOG_LEVEL,
+    setup_logger,
+)
 
 from ..ibm_test_case import IBMTestCase
